+++ src/main/scala/serverciteapp/CiteBinaryImageView.scala
@@ -106,15 +106,23 @@ object CiteBinaryImageView {
 	@dom
 	def imageCollectionsContainer = {
 		<div id="image_imageCollectionsContainer">
-			<h2>Image Collections</h2>
+			<h2><!-- Image Collections --></h2>
 			{
 				<ul>
 					{ for ( ic <- CiteBinaryImageModel.binaryImageCollections) yield {
 						<li>
 							<a
 								onclick={ event: Event => {
-								DataModelController.retrieveObject(None,ic) }
-								}> { s"${ic}" } </a>
+									val mouseEvent = event.asInstanceOf[MouseEvent]
+									if (mouseEvent.metaKey){
+										true
+									} else {	
+										DataModelController.retrieveObject(None,ic) }
+										false
+									}
+								}
+								href={ s"?urn=${ic}" }
+								> { s"${ic}" } </a>
 							<br/>
 								{ ObjectController.labelForCollection(ic) }
 							</li>