+++ src/main/scala/serverciteapp/ObjectView.scala
@@ -375,7 +375,7 @@ def numberInCollection(u:Cite2Urn) = {
 	ObjectModel.labelMap.bind match {
 		case Some(lm) => {
 			val i:Integer = lm.filterKeys( _ ~~ u ).size
-			s"${i} object${if (i > 1) "s"}." 
+			s"${i} object${if (i > 1) "s" else ""}." 
 		}
 		case None => {
 			"No objects."
@@ -389,15 +389,22 @@ def numberInCollection(u:Cite2Urn) = {
 /* General-use functions for making clickable URNs */
 @dom
 def collectionUrnSpan(urn:Cite2Urn) = {
-	<span
+	<a
 	class="app_clickable"
 	onclick={ event: Event => {
+		val mouseEvent = event.asInstanceOf[MouseEvent]
+		if (mouseEvent.metaKey){
+			true
+		} else {
 			ObjectController.insertFirstObjectUrn(urn) 
 			ObjectModel.clearObject
+			false
 		}
-	}>
+	} }
+	href={ s"?urn=${urn}" }
+	>
 	{ urn.toString }
-	</span>
+	</a>
 }
 
 /* For URN properties in objects */