+++ releases.md
@@ -1,5 +1,19 @@
 # Release notes
 
+**1.9.0**: Better handling of links: allows opening URNs in a new window, with a new instance of the App.
+
+**1.8.0**: Adding link to look up any object, or Cite2Urn propert-value in Relations.
+
+**1.7.2**: Minor bug reporting number of objects in a collection.
+
+**1.7.1**: Fixed "race condition" involving image-rois.
+
+**1.7.0**: Accepting a `urn=` request parameter, CTS or CITE2 URN.
+
+**1.6.0**: Commentary implemented.
+
+**1.5.0**: Relations implemented.
+
 **1.4.0**: DSE Data Model implemented. There remain issues (see GitHub) with asynchronous loading of DSE data. 
 
 **1.3.0**: Rebuild CiteCollection handling to push much more work back onto the server.